backend/requirements.txt
django-filter==2.4.0
djoser==2.1.0
drf-extra-fields==3.1.1
pillow==8.3.2
psycopg2-binary==2.8.6
gunicorn==20.1.0